+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # PatchScene CHANGELOG
 
+## [Unreleased]
+
 ## [0.3.5]
 ### Added:
 - new devices and furniture